Remove literate version of user config
Before this commit, `README.org` stored a literate version of the user
configuration in `user-config.el`. To update the latter, I had to update the
README and *tangle* it. This made the README look nice but was more hassle than
the configuration was worth. This commit removes the literate version of the
user configuration.