Update emacs-nw script to address various small issues
This commit lets the script start the Emacs /client/ instead of a full-blown
Emacs instance. This was a regression from an earlier commit. It also removes
the TERM setting that was done before Emacs was started. On my systems, I moved
that responsibility to the shell. Finally, it adds a help to the script.