Show complete output in compilation buffer
As of Emacs 29.1, the compilation buffer can hide the part of a line whose
length exceeds a certain threshold, which it also does by default. This commit
disables that feature.

I've disabled this feature because the pytest progress indicators - a "." for
each test - end up on a single line in a compilation buffer. That line exceeds
that threshold and part of its output is hidden.