Allow pylsp to sort Python imports
In commit 2f2d92c I disabled the use of `isort` to sort the imports, thinking
the Astral ruff server would do that. That's not the case: although the ruff
server has a code action to order imports, it doesn't do this as part of its
formatting.

The pylsp server does not sort imports out-of-the-box, but it has an isort
plugin that does support it. You have to install the Python package for that
plugin, `python-lsp-isort`, and enable it. This commit enables the plugin but
this setting will only have any effect if you also install the plugin.

Emacs lsp-mode assigns a higher prioritiy to the pylsp server (-10 than to the
ruff one (-1), so it should do the isort before the ruff format. This is what we
want.